+++ Assets/Classes/QIEngine/QIEngine(Plugin)/Native C#/Notification.cs
@@ -5,11 +5,11 @@ using UnityEngine;
 
 public class Notification
 {
-    public Action<float> OnConfidenceChanged;
+    public Action<double> OnConfidenceChanged;
     public Action OnQIEngineUpdate;
     public Action OnSelected;
     public Action OnDeselected;
 
     //Internal notification to calculate node confidence
-    public Action<Node, float, bool> GetConfidence;
+    public Action<Node, double, bool> GetConfidence;
 }
